Created scoreboard component folder with all the classes
