+++ olympics-website/app/api/scoreboard/route.ts
@@ -1,0 +1,1 @@
+import { NextResponse } from "next/server";
