+++ olympics-website/components/scoreboard/Team.tsx
@@ -1,0 +1,6 @@
+export default interface Team {
+  id: number;
+  name: string;
+  scores: Record<string, number>;
+  score: number;
+}
